Moved assessment-related files to component folder and fixed import paths; made requested change to level assignment
